Moved dummy files to dummy folder
